DataBase Conn

+++ skuhub/src/main/java/org/skuhub/skuhub/repository/sample/SampleRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SampleRepository extends JpaRepository<SampleJpaEntity, Long>, SampleDslRepository {
+public interface SampleRepository extends JpaRepository<SampleJpaEntity, Integer>, SampleDslRepository {
 }
